mapping-related stuff migrated from Parameters subnamespace to Mapping
