+++ Revit_Core_Engine/Convert/Revit/FromRevit/Parameter.cs
@@ -21,6 +21,7 @@
  */
 
 using Autodesk.Revit.DB;
+using BH.oM.Adapters.Revit.Mapping;
 using BH.oM.Adapters.Revit.Parameters;
 using System.Collections.Generic;
