+++ Revit_Engine/Create/Generic/ParameterMap.cs
@@ -20,7 +20,7 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using BH.oM.Adapters.Revit.Parameters;
+using BH.oM.Adapters.Revit.Mapping;
 using BH.oM.Reflection.Attributes;
 using System;
 using System.Collections.Generic;
